flaskapp/requirements.txt
Flask==0.11.1
gunicorn==19.6.0
Pyrebase4==4.4.3